Turn off window decorations when going fullscreen under X
Older versions of gnome-shell don't seem to hide the top panel when a
window goes fullscreen unless the decorations are also turned off for
that window. Pinpoint already has code to turn off the decorations but
it is only used as a workaround if the window is maximised or
something. This replaces that workaround and makes it always use
clutter_actor_set_fullscreen but it additionally reuses that code to
turn off the decorations.